Create or modify a project's DESCRIPTION file

A simple wrapper around usethis::use_description() to write a DESCRIPTION file in the root project directory. This file can be used to record project metadata, including package version information from the current project library.

description(fields = list(), library = NULL, snapshot = NULL)

Arguments

fields

named list of DESCRIPTION fields and their values.

library

paths to package libraries. if NULL, .libPaths() will be used.

snapshot

character. path to a renv or Require snapshot file.

Value

NULL invisibly. Invoked for side effect of writing a DESCRIPTION file to the project directory, and printing the resulting DESCRIPTION file to screen.

Note

fields Imports and Remotes will be automatically populated based on snapshot (or, if snapshot = NULL, the currently installed packages).

Examples

tmpdir <- file.path(tempdir(), "example_project") |>
  Require::checkPath(create = TRUE)
cwd <- setwd(tmpdir)
opts <- options(workflowtools.useRequire = TRUE)

description(fields = list(
  Title = "My Project: It Does Cool Stuff",
  Description = paste(
    "My project does so many cool things.",
    "It's so useful for all the things."
  ),
  `Authors@R` = "c(
    person('First', 'Last', , 'email@email.com', role = 'aut')
  )",
  Version = "0.0.1",
  Language = "en-CA",
  License = "MIT",
  Depends = paste0("R (== 4.3)", collapse = ",\n    ")
))
#> package version file saved in /tmp/Rtmpzs2eJz/pkgsnapshot_Require_6c067f84e78.csv
#> ✔ Setting active project to '/tmp/Rtmpzs2eJz/example_project'
#> ✔ Writing 'DESCRIPTION'
#> Type: project
#> Package: example_project
#> Title: My Project: It Does Cool Stuff
#> Version: 0.0.1
#> Authors@R (parsed):
#>     * First Last <email@email.com> [aut]
#> Description: My project does so many cool things. It's so useful
#>     for all the things.
#> License: MIT
#> Depends:
#>     R (== 4.3)
#> Imports:
#>     DBI (== 1.2.2),
#>     R6 (== 2.5.1),
#>     Rcpp (== 1.0.12),
#>     Require (== 0.3.1),
#>     askpass (== 1.2.0),
#>     base64enc (== 0.1-3),
#>     brio (== 1.1.5),
#>     bslib (== 0.7.0),
#>     cachem (== 1.0.8),
#>     callr (== 3.7.6),
#>     classInt (== 0.4-10),
#>     cli (== 3.6.2),
#>     clipr (== 0.8.0),
#>     covr (== 3.6.4),
#>     cpp11 (== 0.4.7),
#>     crayon (== 1.5.2),
#>     credentials (== 2.0.1),
#>     curl (== 5.2.1),
#>     data.table (== 1.15.4),
#>     desc (== 1.4.3),
#>     details (== 0.3.0),
#>     diffobj (== 0.3.5),
#>     digest (== 0.6.35),
#>     downlit (== 0.4.3),
#>     dplyr (== 1.1.4),
#>     e1071 (== 1.7-14),
#>     evaluate (== 0.23),
#>     fansi (== 1.0.6),
#>     fastmap (== 1.1.1),
#>     fontawesome (== 0.5.2),
#>     fs (== 1.6.3),
#>     furrr (== 0.3.1),
#>     future (== 1.33.2),
#>     gargle (== 1.5.2),
#>     generics (== 0.1.3),
#>     gert (== 2.0.1),
#>     gh (== 1.4.1),
#>     gitcreds (== 0.1.2),
#>     globals (== 0.16.3),
#>     glue (== 1.7.0),
#>     googledrive (== 2.1.1),
#>     highr (== 0.10),
#>     htmltools (== 0.5.8.1),
#>     httr (== 1.4.7),
#>     httr2 (== 1.0.1),
#>     ini (== 0.3.1),
#>     jquerylib (== 0.1.4),
#>     jsonlite (== 1.8.8),
#>     knitr (== 1.46),
#>     lazyeval (== 0.2.2),
#>     lifecycle (== 1.0.4),
#>     listenv (== 0.9.1),
#>     magrittr (== 2.0.3),
#>     memoise (== 2.0.1),
#>     mime (== 0.12),
#>     openssl (== 2.1.2),
#>     pak (== 0.7.2),
#>     parallelly (== 1.37.1),
#>     pillar (== 1.9.0),
#>     pkgbuild (== 1.4.4),
#>     pkgconfig (== 2.0.3),
#>     pkgdown (== 2.0.9),
#>     pkgload (== 1.3.4),
#>     png (== 0.1-8),
#>     praise (== 1.0.0),
#>     processx (== 3.8.4),
#>     proxy (== 0.4-27),
#>     ps (== 1.7.6),
#>     purrr (== 1.0.2),
#>     ragg (== 1.3.0),
#>     rappdirs (== 0.3.3),
#>     rematch2 (== 2.1.2),
#>     remotes (== 2.5.0),
#>     renv (== 1.0.7),
#>     rex (== 1.2.1),
#>     rlang (== 1.1.3),
#>     rmarkdown (== 2.26),
#>     rprojroot (== 2.0.4),
#>     rstudioapi (== 0.16.0),
#>     s2 (== 1.1.6),
#>     sass (== 0.4.9),
#>     sessioninfo (== 1.2.2),
#>     sf (== 1.0-16),
#>     sys (== 3.4.2),
#>     systemfonts (== 1.0.6),
#>     testthat (== 3.2.1.1),
#>     textshaping (== 0.3.7),
#>     tibble (== 3.2.1),
#>     tidyselect (== 1.2.1),
#>     tinytex (== 0.50),
#>     units (== 0.8-5),
#>     usethis (== 2.2.3),
#>     utf8 (== 1.2.4),
#>     uuid (== 1.2-0),
#>     vctrs (== 0.6.5),
#>     waldo (== 0.5.2),
#>     whisker (== 0.4.1),
#>     withr (== 3.0.0),
#>     wk (== 0.9.1),
#>     workflowtools (== 0.0.3),
#>     xfun (== 0.43),
#>     xml2 (== 1.3.6),
#>     yaml (== 2.3.8),
#>     zip (== 2.3.1)
#> Remotes:
#>     /@
#> Encoding: UTF-8
#> Language: en-CA

setwd(cwd)
unlink(tmpdir, recursive = TRUE)
options(opts)